Продолжение оформления архитектуры приложения

+++ client/inc/TransformComponent.hpp
@@ -5,7 +5,7 @@
 #ifndef TRANSFORM_COMPONENT_TYPE_HEADER
 #define TRANSFORM_COMPONENT_TYPE_HEADER
 
-#include "Vec3i.hpp"
+#include "Vec3d.hpp"
 
 #include "Component.hpp"
 
@@ -13,9 +13,9 @@ namespace tppo{
     class TransformComponent : Component {
         
     protected:
-        Vec3i position;
-        Vec3i size;
-        Vec3i rotation;
+        Vec3d position;
+        Vec3d size;
+        Vec3d rotation;
         
     public:
